HACK: futex: don't optimize futex_detect_cmpxchg() on ARM64
On ARM64, clang may propagate the uaddr = NULL constant all the way from
futex_detect_cmpxchg() to futex_atomic_cmpxchg_inatomic(), and try to
fetch from [xzr] (which there's no way to assemble).

Prevent this by disabling optimizations in futex_detect_cmpxchg(). This
is ugly, but it only impacts a function that runs once at boot.

This is a workaround, b:37425259 might lead to a proper solution.

BUG=chromium:702741, b:37425259
TEST=build and boot on gru

Change-Id: I46276159a97d62c03a0719620ebc69642df87fb3
Signed-off-by: Greg Hackmann <ghackmann@google.com>
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/528533
Reviewed-by: Guenter Roeck <groeck@chromium.org>