CHROMIUM: bluetooth: update advertising parameters for single advertisement
Currently, the advertisement scheduling algorithm does not update
the advertising data and parameters on the controller if there is
only one advertisement. This saves effort when the advertisement
data and parameters are not changed.

Since new methods have been provided to change advertising intervals
and duration, it is needed to have a flag to indicate the change
and update the controller accordingly when the flag is set true.

BUG=chromium:673499
TEST=Use the "nRF Connect" android app to verify.
Single advertisement:
- register a single advertisement,
- change the advertising interval, and
- check if the new interval takes effect immediately.
Also verify multiple advertisements work correctly as well.

Change-Id: I058e4c7a9feeee691f507903037e581e9eb2978d
Reviewed-on: https://chromium-review.googlesource.com/425558
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Miao-chen Chou <mcchou@chromium.org>
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Rahul Chaturvedi <rkc@chromium.org>
Reviewed-by: Miao-chen Chou <mcchou@chromium.org>