CHROMIUM: nouveau: Fix an use-after-free nvkm_handle issue
If we send a nouveau client delete IOCTL to kernel, nouveau
driver forgets to set the "nvkm_handle" pointer to NULL.
This brings an issue that, if nouveau starts runtime suspend
after that, kernel NULL pointer crash happens.

This patch fixes this by setting the pointer to NULL when deleting
nvkm_handle of nvkm_client, besides adds checking in every places
which dereferences this handle.

Be noticed that ideally this should be fixed in upstream nouveau.
It "seems" that upstream nouveau has fixed this but due to the
big divergence between A44 nouveau and upstream's, we don't want
to bother backporting lots of upstream patches(also fixing many
conflicts) so this simple version fix is created.

BUG=chrome-os-partner:61490, b:36117306
TEST=System boots OK / Youtube works/ POC app doesn't crash kernel

Change-Id: Iccaad7f35a0a527e942d8ad73a3bc547b54326a2
Signed-off-by: Mark Zhang <markz@nvidia.com>
(cherry picked from commit 212850f06fa03defcbdee18b7c06fdf4b4b93ca2)
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/454997
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>