BACKPORT: drm/i915: Simplify testing for am-I-the-kernel-context?
The kernel context (dev_priv->kernel_context) is unique in that it is
not associated with any user filp - it is the only one with
ctx->file_priv == NULL. This is a simpler test than comparing it against
dev_priv->kernel_context which involves some pointer dancing.

In checking that this is true, we notice that the gvt context is
allocating itself a i915_hw_ppgtt it doesn't use and not flagging that
its file_priv should be invalid.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Link: http://patchwork.freedesktop.org/patch/msgid/20170106152013.24684-5-chris@chris-wilson.co.uk
(cherry picked from commit 984ff29f74c0c130b43f0c5b0fe0fbca5de0fddc)
drm-intel-next-2017-01-09
Signed-off-by: Tarun <tarun.vyas@intel.com>

BUG=b:36446508
TEST=No regressions in KBL and APL based devices

 Conflicts: [Tarun: Avoid "c8c3579 drm/i915: Introduce GVT context creation API".
 		GVT stuff isn't needed]
	drivers/gpu/drm/i915/i915_gem_context.c

Change-Id: Id84e27c1b451d48c43c9d05f81ad778bd2237fb7