UPSTREAM: r8152: fix rx issue for runtime suspend
Pause the rx and make sure the rx fifo is empty when the autosuspend
occurs.

If the rx data comes when the driver is canceling the rx urb, the host
controller would stop getting the data from the device and continue
it after next rx urb is submitted. That is, one continuing data is
split into two different urb buffers. That let the driver take the
data as a rx descriptor, and unexpected behavior happens.

Signed-off-by: Hayes Wang <hayeswang@realtek.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 75dc692eda114cb234a46cb11893a9c3ea520934)

BUG=chromium:768542
TEST=netperf, repeated reboot, and network_EthernetStressPlug on sentry

Change-Id: Ic906b80b3cc0dc16aca2dec7314db1e67dc19fb5
Signed-off-by: Grant Grundler <grundler@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/876806
Reviewed-by: Guenter Roeck <groeck@chromium.org>