CHROMIUM: drm/rockchip: vop: Do not update primary plane without framebuffer
For some reason vop_update_primary_plane() might get called without
a framebuffer attached to primary plane. To avoid a crash from a NULL
pointer dereference let's just bail out from the function. This should
be safe, as initially after resume we should have all planes disabled
in hardware.

We should investigate the root cause, but since this seems to be a quite
common crash on production devices and it does not seem to reproduce
locally, we can work it around by this change.

BUG=chromium:595281
TEST=suspend_stress_test

Change-Id: I30c766869a627ce5a20d4e91e8e4d8b035631134
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/428772
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>