CHROMIUM: rk3399: Disable ddr clk timeout on suspend
This disables the timeout for the ddr clk when the governor is
suspended or stopped. This makes sure that the suspend frequency is
set. It also makes sure that the userspace governor will be able to
change the frequency without failing.

BUG=chrome-os-partner:62193
TEST=boot kevin
plug and unplug external monitor
make sure that dpll is always 928000000 in
/sys/kernel/debug/clk/clk_summary

Change-Id: I9c53135ea9dcbbcf899d2d6134c8a6e998e2bc6a
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/431203
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
(cherry picked from commit e2cc20adc4fd51b1013ac687c304c6545d70041b)
Reviewed-on: https://chromium-review.googlesource.com/433099
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>