CHROMIUM: rk3399: dmc: fix enable/disable issue
The block and unblock functions for the dmc did not check the number
of notifiers registered. This could result in unblock resuming the dmc
when there were 2 or more notifiers, which is an issue. This adds in
the additional check for both the block and unblock functions.

BUG=chrome-os-partner:54651
TEST=boot kevin and plug/unplug and external monitor

Change-Id: Id390eac8d39bf84dc8bae56bdc702992c7650eec
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/431205
Reviewed-by: Douglas Anderson <dianders@chromium.org>
(cherry picked from commit 599e97d6eacf8e677e273fadee1b03dd2c1c0e44)
Reviewed-on: https://chromium-review.googlesource.com/433103
Tested-by: Douglas Anderson <dianders@chromium.org>