CHROMIUM: drm: bridge/analogix: fix timeout of video streamclk config
The STRM_VALID bit in register ANALOGIX_DP_SYS_CTL_3 may be unstable,
so we may hit the error log "Timeout of video streamclk ok" since
checked this unstable bit.
In fact, we can go continue and the streamclk is ok if we wait enough time,
it does no effect on display.
Let's change this error to warn.

BUG=chrome-os-partner:62222
TEST=boot on kevin

Change-Id: I4e0ad019eaf733b645efb26c341868272ffdd4b6
Signed-off-by: zain wang <wzz@rock-chips.com>
Reviewed-on: https://chromium-review.googlesource.com/431257
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>