CHROMIUM: iwl7000: mvm: fix rogue AP detection for MQ
In MQ, the tpt we demand for declaring that the AP failed
AP detection is more than 22 percent of the phy rate. So if
the traffic is not more than 22 percent of the phy rate, the
AP can't fail AP detection, and then there is no reason for
disconnecting from it and cancelling U-APSD.

Flip the inequation mark so it would reflect this logic.

Signed-off-by: Golan Ben-Ami <golan.ben.ami@intel.com>
iwl7000-tree: 28907937da0357c1c0809b50caa559b2729fdb34