UPSTREAM: arm/arm64: KVM: Enforce unconditional flush to PoC when mapping to stage-2
commit 8f36ebaf21fdae99c091c67e8b6fab33969f2667 upstream.

When we fault in a page, we flush it to the PoC (Point of Coherency)
if the faulting vcpu has its own caches off, so that it can observe
the page we just brought it.

But if the vcpu has its caches on, we skip that step. Bad things
happen when *another* vcpu tries to access that page with its own
caches disabled. At that point, there is no garantee that the
data has made it to the PoC, and we access stale data.

The obvious fix is to always flush to PoC when a page is faulted
in, no matter what the state of the vcpu is.

Fixes: 2d58b733c876 ("arm64: KVM: force cache clean on page fault when caches are off")
Reviewed-by: Christoffer Dall <christoffer.dall@linaro.org>
Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Salman Qazi <sqazi@chromium.org>

BUG=chromium:909793
TEST=See 909793 for instructions on reproducing the bug.  Briefly: run
'dd if=/dev/zero of=foo.bin BS=4M count=1K' on btrfs in crosvm on hana on a big
core.  Prior to this change, this produces undefined instructions
exception.  After this patch, it doesn't.

(cherry picked from commit e83a28c521a6a9df75c031166cf18b5a5202ecdb in
linux-stable)

Change-Id: Ibf1a0613f5c3dfd5d34cde65ad612b9c060018c1
Reviewed-on: https://chromium-review.googlesource.com/1601659
Commit-Ready: Salman Qazi <sqazi@google.com>
Tested-by: Salman Qazi <sqazi@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
(cherry picked from commit ef5f1d827f7c235e13cc6e1189a6f355d724ba8e)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1618206
Commit-Queue: Stephen Barber <smbarber@chromium.org>
Tested-by: Stephen Barber <smbarber@chromium.org>