CHROMIUM: arm64: dts: rockchip: set PCLK_EDP to 25M (not 24M)
Previously we landed a chance to set the EDP clock to "24 MHz" to avoid
aux channel errors.  Apparently this causes problems on some units.

A few notes:
* 24 MHz actually ended up getting 23.76 MHz since you can't make 24
  MHz by diving 800 MHz (CPLL) by an integer.  23.76 was the best the
  kernel could do by dividing 594 MHz (GPLL) by an integer (/ 25).
* In coreboot we chose 25 MHz (http://crosreview.com/429690), which is
  an integral divider.

On units that failed with 23.76 MHz it can be seen that they work with
25 MHz.  We'll switch to that because:

* It's definitely far away from 100 MHz, which was also apparently
  causing problems on some units.
* It matches what we did for coreboot.
* It seems to work.

If we find that 25 MHz _still_ doesn't work on some units, we can try
something that more "in the middle", but hopefully we can just be
happy at 25 MHz.

BUG=chrome-os-partner:60130
TEST=Bad machine no longer reports errors.

Change-Id: I39eddf8d5d1c3916f8da987867a8011f1c6839e7
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/433919