CHROMIUM: arm64: dts: rockchip: Set GPLL to 600 MHz, not 594 MHz on rk3399-gru
Apparently there is a hardware errata in our eDP block where it
sometimes has trouble syncing the PHY an PCLK clocks.  To avoid this
problem we really want them to both be at exactly 24 MHz.

The eDP PCLK can be parented off of CPLL and GPLL, so we need to make
one of these a multiple of 24 MHz.  Chaning GPLL from 594 MHz to 600
MHz seems line a sane thing to do.

One might ask why GPLL was originally 594 MHz instead of 600 MHz.
This was primarily to allow the SoC to make these pixel clocks:
  74.25
  148.5
  297
  594
...without needing to use the variable PLL.  The idea is that if you
have two displays (like HDMI and eDP) that you could use GPLL as the
source for one and the variable PLL as the source for the other.

Unfortunately the idea of 594 MHz being super useful never really
worked out, at least in all cases I've seen.  Usually you don't just
want the pixels clocks that are multiples of 74.25 but you also want
other pixel clocks (so you can support arbitrary DVI resolutions).  So
really the only way to support two displays is for one of them to be
the "builtin" display and find some sort of pixel clock that can be
made from CPLL or GPLL.

An example would be http://crosreview.com/373047 where pixel clocks
were adjusted.  Part of that patch landed for rk3399-kevin's panel.
Nothing similar ever landed for gru, so presumably gru's panel isn't
quite 60 Hz right now.

NOTE: for now we're only editing files relating to rk3399-gru.  At
some point an upstream solution will be needed for this.

BUG=chrome-os-partner:60130
TEST=Stress the heck out of everything

Change-Id: Ic0f7e149a6f4787cfcfb1cd9e0ea07e4e4af93c9
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/434140