CHROMIUM: iwl7000: mvm: move number of words calculation in MFUART to a variable
The type of the value in the offset calculation was different for
32-bit and 64-bit architectures, so the format string would cause a
warning in 32-bit archs.  To avoid this, and to make the code look a
bit cleaner, move the number of words calculation to an integer
variable and use that in the loop.

Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: e6cd611484ed2556f458a44d34d580dab071ff13