BACKPORT: PM / devfreq: Fix available_governor sysfs
The devfreq using passive governor is not able to change the governor.
So, the user can not change the governor through 'available_governor' sysfs
entry. Also, the devfreq which don't use the passive governor is not able to
change to 'passive' governor on the fly.

Fixes: 996133119f57 ("PM / devfreq: Add new passive governor")
Cc: stable@vger.kernel.org
Signed-off-by: Chanwoo Choi <cw00.choi@samsung.com>
Signed-off-by: MyungJoo Ham <myungjoo.ham@samsung.com>
(cherry picked from commit bcf23c79c4e46130701370af4383b61a3cba755c)

Conflicts:
  drivers/devfreq/governor_passive.c
    The passive governor does not exist in v4.4. We still want this change
    since it helps to avoid conflicts when applying other changes.

BUG=b:74321682, b:73050145
TEST=build for scarlet and boot

Change-Id: I27c03be96eae08dac204db1eeee514146c289e4f
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1067726
Reviewed-by: Brian Norris <briannorris@chromium.org>