CHROMIUM: Bluetooth: Add BT splitter
BT splitter is a simple BT device virtualization layer that allows
bluez to share the chip with a userspace stack. The userspace stack
handles all LE and bluez - EDR. This code hides LE capabilities of the
chip from bluez, and thus no more changes to bluez are likely
to be needed. The userspace stack is built aware of this splitter
and acts accordingly (not using EDR), though this is not necessarily
required, as the userspace stack can be shown an LE-only chip too.
We choose not to do that as we control that stack already and there is
thus no need for this. There are improvements that can be made (like
seeing if there are separate LE buffers before stealing EDR bufers),
which we'll address soon.

The way userspace connects to this:
 * open /dev/hci_splitter and treat it asn an hci device
 * when opened a RESET command complete packet will be read
 * it will have the "number of comands" value set to 0 or 1
   * 0 if chip i snot yet ready (and stack will wait as per spec)
   * 1 if chip is ready
 * if userspace dies, chip is reset (as no state exosts to
   restore it back to normalcy).
 * if bluez resets chip or otherwise shuts it down, another
   command complete event is sent as a reset (as chip is indeed
   reset) with "number of commands" value set to zero

TEST=verified bluez remains functional and does not see any LE features with splitter on
BUG=35775020

Change-Id: I2ad520b26bf7a35dc98ed2548c936734afd98f39
Signed-off-by: Dmitry Grinberg <dmitrygr@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1558599
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Sonny Sasaka <sonnysasaka@chromium.org>
Reviewed-by: Miao-chen Chou <mcchou@chromium.org>